+++ Tests/BindgenXcodePluginFixture/Sources/AppDelegate.swift
@@ -8,8 +8,8 @@ final class AppDelegate: UIResponder, UIApplicationDelegate {
         didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
     ) -> Bool {
         _ = ManagementCanister.self
-        _ = ManagementCanisterStatusArgs.self
-        _ = ManagementCanisterStatus.self
+        _ = ManagementCanisterInfoArgs.self
+        _ = ManagementCanisterInfo.self
         return true
     }
 }